requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-trusted-host --no-index requirements.in
#
cachetools==4.1.0         # via google-auth
certifi==2020.6.20        # via kubernetes, requests
chardet==3.0.4            # via requests
deepdiff==3.3.0           # via -r requirements.in
google-auth==1.18.0       # via kubernetes
idna==2.9                 # via requests
importlib-metadata==1.6.1  # via jsonpickle
jsonpickle==1.4.1         # via deepdiff
kubernetes==11.0.0        # via -r requirements.in
oauthlib==3.1.0           # via requests-oauthlib
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
python-dateutil==2.8.1    # via kubernetes
pyyaml==5.3.1             # via -r requirements.in, kubernetes
requests-oauthlib==1.3.0  # via kubernetes
requests==2.24.0          # via -r requirements.in, kubernetes, requests-oauthlib
rsa==4.6                  # via google-auth
six==1.15.0               # via google-auth, kubernetes, python-dateutil, websocket-client
urllib3==1.25.9           # via kubernetes, requests
websocket-client==0.57.0  # via kubernetes
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
